refactor: Move old helpers to History folder
