+++ src/views/Predictions/components/History/PnlTab/PnlTab.tsx
@@ -3,12 +3,12 @@ import styled from 'styled-components'
 import { useWeb3React } from '@web3-react/core'
 import { Box, Flex, Heading, Text, Button, Link, OpenNewIcon } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
+import { getRoundResult, Result } from 'state/predictions/helpers'
 import { getBscScanAddressUrl } from 'utils/bscscan'
 import store from 'state'
 import { useGetCurrentEpoch, usePriceBnbBusd } from 'state/hooks'
 import { Bet, BetPosition } from 'state/types'
-import { formatBnb, getMultiplier, getNetPayout } from 'views/Predictions/helpers'
-import { getRoundResult, Result } from 'state/predictions/helpers'
+import { formatBnb, getMultiplier, getNetPayout } from '../helpers'
 import PnlChart from './PnlChart'
 import SummaryRow from './SummaryRow'
