+++ src/views/Predictions/components/RoundResult/RoundResultHistory.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { BoxProps, Flex, Text } from '@pancakeswap/uikit'
 import { BetPosition, Round } from 'state/types'
 import { useTranslation } from 'contexts/Localization'
-import { formatUsd } from '../../helpers'
+import { formatUsd } from '../History/helpers'
 import PositionTag from '../PositionTag'
 import { LockPriceHistoryRow, PrizePoolHistoryRow, RoundResultBox } from './styles'
