+++ src/views/Predictions/components/RoundResult/styles.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import styled, { DefaultTheme } from 'styled-components'
 import { Box, Flex, FlexProps, Text } from '@pancakeswap/uikit'
-import { formatBnb, formatUsd, formatUsdv2, formatBnbv2 } from 'views/Predictions/helpers'
 import { useTranslation } from 'contexts/Localization'
 import { BetPosition, NodeRound, Round } from 'state/types'
+import { formatUsdv2, formatBnbv2 } from '../../helpers'
+import { formatBnb, formatUsd } from '../History/helpers'
 
 // PrizePoolRow
 interface PrizePoolRowProps extends FlexProps {
