+++ src/views/Predictions/helpers.ts
@@ -1,32 +1,15 @@
-import BigNumber from 'bignumber.js'
 import { ethers } from 'ethers'
-import { Bet, BetPosition, NodeLedger, NodeRound } from 'state/types'
-import { formatBigNumber, formatBigNumberToFixed, formatNumber, getBalanceAmount } from 'utils/formatBalance'
+import { BetPosition, NodeLedger, NodeRound } from 'state/types'
+import { formatBigNumber, formatBigNumberToFixed } from 'utils/formatBalance'
 import getTimePeriods from 'utils/getTimePeriods'
-
-export const getBnbAmount = (bnbBn: BigNumber) => {
-  return getBalanceAmount(bnbBn, 18)
-}
-
-export const formatUsd = (usd: number) => {
-  return `$${formatNumber(usd || 0, 3, 3)}`
-}
 
 export const formatUsdv2 = (usd: ethers.BigNumber) => {
   return `$${formatBigNumberToFixed(usd, 3, 8)}`
 }
 
-export const formatBnb = (bnb: number) => {
-  return bnb ? bnb.toLocaleString(undefined, { minimumFractionDigits: 3, maximumFractionDigits: 3 }) : '0'
-}
-
 export const formatBnbv2 = (bnb: ethers.BigNumber) => {
   const value = bnb || ethers.BigNumber.from(0)
   return formatBigNumberToFixed(value, 4)
-}
-
-export const formatBnbFromBigNumber = (bnbBn: BigNumber) => {
-  return formatBnb(getBnbAmount(bnbBn).toNumber())
 }
 
 export const padTime = (num: number) => num.toString().padStart(2, '0')
@@ -40,36 +23,6 @@ export const formatRoundTime = (secondsBetweenBlocks: number) => {
   }
 
   return minutesSeconds
-}
-
-export const getMultiplier = (total: number, amount: number) => {
-  if (total === 0 || amount === 0) {
-    return 0
-  }
-
-  return total / amount
-}
-
-/**
- * Calculates the total payout given a bet
- */
-export const getPayout = (bet: Bet, rewardRate = 1) => {
-  if (!bet || !bet.round) {
-    return 0
-  }
-
-  const { bullAmount, bearAmount, totalAmount } = bet.round
-  const multiplier = getMultiplier(totalAmount, bet.position === BetPosition.BULL ? bullAmount : bearAmount)
-  return bet.amount * multiplier * rewardRate
-}
-
-export const getNetPayout = (bet: Bet, rewardRate = 1): number => {
-  if (!bet || !bet.round) {
-    return 0
-  }
-
-  const payout = getPayout(bet, rewardRate)
-  return payout - bet.amount
 }
 
 export const getHasRoundFailed = (round: NodeRound, blockNumber: number) => {
@@ -111,13 +64,14 @@ export const getPayoutv2 = (ledger: NodeLedger, round: NodeRound, rewardRate = 1
     return ethers.BigNumber.from(0)
   }
 
-  const { bullAmount, bearAmount, totalAmount } = round
-  const multiplier = getMultiplierv2(totalAmount, ledger.position === BetPosition.BULL ? bullAmount : bearAmount)
+  const { bullAmount, bearAmount, rewardAmount } = round
+  const { amount, position } = ledger
+  const multiplier = getMultiplierv2(rewardAmount, position === BetPosition.BULL ? bullAmount : bearAmount)
 
   const multiplierUnits = ethers.utils.parseUnits(multiplier)
   const rewardRateUnits = ethers.utils.parseUnits(rewardRate.toString())
 
-  return bearAmount.mul(multiplierUnits).mul(rewardRateUnits)
+  return amount.mul(multiplierUnits).mul(rewardRateUnits)
 }
 
 export const getNetPayoutv2 = (ledger: NodeLedger, round: NodeRound, rewardRate = 1) => {
